+++ source/Messaging.Api/MasterDataReceivers/CustomerMasterDataResponseListener.cs
@@ -19,8 +19,8 @@ using Energinet.DataHub.EnergySupplying.RequestResponse.Requests;
 using Messaging.Api.Configuration.Middleware;
 using Messaging.Application.Configuration.Commands;
 using Messaging.Application.MasterData;
-using Messaging.Application.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Messaging.Application.Transactions.MoveIn.MasterDataDelivery;
+using Messaging.Domain.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Messaging.Infrastructure.Configuration.InternalCommands;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
