+++ source/Messaging.Application/Transactions/MoveIn/MasterDataDelivery/CustomerMasterDataMessageFactory.cs
@@ -16,10 +16,10 @@ using System;
 using System.Threading.Tasks;
 using Messaging.Application.Configuration;
 using Messaging.Application.OutgoingMessages;
-using Messaging.Application.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Messaging.Application.OutgoingMessages.Common;
 using Messaging.Domain.Actors;
 using Messaging.Domain.OutgoingMessages;
+using Messaging.Domain.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Messaging.Domain.Transactions;
 using Messaging.Domain.Transactions.MoveIn;
