+++ source/Messaging.Application/Transactions/MoveIn/Notifications/MoveInNotifications.cs
@@ -16,9 +16,9 @@ using System;
 using Messaging.Application.Configuration;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Application.OutgoingMessages.Common;
-using Messaging.Application.OutgoingMessages.GenericNotification;
 using Messaging.Domain.Actors;
 using Messaging.Domain.OutgoingMessages;
+using Messaging.Domain.OutgoingMessages.GenericNotification;
 using Messaging.Domain.Transactions;
 using Messaging.Domain.Transactions.MoveIn;
 using NodaTime;
