+++ source/Messaging.Domain/OutgoingMessages/CharacteristicsOfACustomerAtAnAp/MarketActivityRecord.cs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Collections.Generic;
 using NodaTime;
 
-namespace Messaging.Application.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
+namespace Messaging.Domain.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 
 public record MarketActivityRecord(string Id, string OriginalTransactionId, Instant ValidityStart, MarketEvaluationPoint MarketEvaluationPoint);
