+++ source/Messaging.Domain/OutgoingMessages/GenericNotification/MarketActivityRecord.cs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using NodaTime;
 
-namespace Messaging.Application.OutgoingMessages.GenericNotification
+namespace Messaging.Domain.OutgoingMessages.GenericNotification
 {
     public class MarketActivityRecord
     {
