+++ source/Messaging.IntegrationTests/Application/Transactions/MoveIn/SampleData.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using Messaging.Application.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
+using Messaging.Domain.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Messaging.IntegrationTests.Factories;
 using NodaTime;
 using NodaTime.Extensions;
