+++ source/Messaging.Tests/Infrastructure/OutgoingMessages/GenericNotification/GenericNotificationDocumentWriterTests.cs
@@ -21,9 +21,9 @@ using System.Xml.Linq;
 using System.Xml.Schema;
 using Messaging.Application.Configuration;
 using Messaging.Application.OutgoingMessages.Common;
-using Messaging.Application.OutgoingMessages.GenericNotification;
 using Messaging.Application.SchemaStore;
 using Messaging.Domain.OutgoingMessages;
+using Messaging.Domain.OutgoingMessages.GenericNotification;
 using Messaging.Infrastructure.Common;
 using Messaging.Infrastructure.Configuration;
 using Messaging.Infrastructure.Configuration.Serialization;
